+++ packages/renderer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-renderer
 
+## 0.1.10
+
 ## 0.1.9
 
 ## 0.1.8